continue-learn/src/run1.py
from flask import Flask
from flask_mysqldb import MySQL

app = Flask(__name__)

app.config['MYSQL_USER'] = 'cuong'
app.config['MYSQL_PASSWORD'] = 'password'
app.config['MYSQL_DB'] = 'PhanMemKhaoThi'
app.config['MYSQL_CURSORCLASS'] = 'DictCursor'
app.config['MYSQL_HOST'] = '192.168.40.228'
app.config['MYSQL_PORT'] = '3306'
mysql = MySQL(app)

@app.route('/')
def users():
    cur = mysql.connection.cursor()
    cur.execute('''SELECT user, host FROM mysql.user''')
    rv = cur.fetchall()
    return str(rv)

if __name__ == '__main__':
    app.run(debug=True)